requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ajax-selects==2.1.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2